Fix Form 1040 PDF field mapping for 2025 IRS form
- Personal info: names start at f1_04 (f1_03 is fiscal-year row)
- Income: f1_54 is 1h description text; amounts cascade from f1_55; taxable income moves to f1_75
- Add NeedAppearances=True so Poppler regenerates appearances from /V values, fixing blank page 2 fields
- Fix EIC handling in tax_service: separate refundable EIC from non-refundable credits
- Use BigInteger for financial cent columns to prevent overflow

+++ backend/app/models/tax_return.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, Enum, Text
+from sqlalchemy import Column, Integer, BigInteger, String, Float, ForeignKey, DateTime, Enum, Text
 from sqlalchemy.sql import func
 from app.database import Base
 import enum
@@ -29,10 +29,10 @@ class TaxReturn(Base):
     status = Column(Enum(ReturnStatus), default=ReturnStatus.draft)
 
     # Financial fields (stored as cents to avoid float issues)
-    estimated_tax_cents = Column(Integer, default=0)
-    tax_owed_cents = Column(Integer, default=0)
-    refund_amount_cents = Column(Integer, default=0)
-    total_income_cents = Column(Integer, default=0)
+    estimated_tax_cents = Column(BigInteger, default=0)
+    tax_owed_cents = Column(BigInteger, default=0)
+    refund_amount_cents = Column(BigInteger, default=0)
+    total_income_cents = Column(BigInteger, default=0)
 
     # IRS e-file tracking
     submission_id = Column(String)
